@@ -5,45 +5,38 @@
     "ng": "ng",
     "start": "ng serve",
     "build": "ng build",
-    "lint": "ng lint",
     "extract-translations": "ngx-translate-extract --input ./src --output ./src/assets/i18n/en.json ./src/assets/i18n/de.json --clean --sort --format namespaced-json --marker _",
-    "deploy": "ng build --prod --baseHref=/frontend/ && cpx ./dist/ida-frontend/**/* ../app/server/src/main/resources/static/frontend/"
+    "deploy": "ng build --configuration=production --baseHref=/frontend/ && cpx ./dist/ida-frontend/**/* ../app/server/src/main/resources/static/frontend/"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~7.2.0",
-    "@angular/common": "~7.2.0",
-    "@angular/compiler": "~7.2.0",
-    "@angular/core": "~7.2.0",
-    "@angular/forms": "~7.2.0",
-    "@angular/platform-browser": "~7.2.0",
-    "@angular/platform-browser-dynamic": "~7.2.0",
-    "@angular/router": "~7.2.0",
-    "@ng-bootstrap/ng-bootstrap": "^4.1.0",
-    "@ngx-translate/core": "^11.0.1",
-    "@ngx-translate/http-loader": "^4.0.0",
-    "@stomp/stompjs": "^5.2.0",
-    "core-js": "^2.5.4",
+    "@angular/animations": "^13.3.11",
+    "@angular/common": "^13.3.11",
+    "@angular/compiler": "^13.3.11",
+    "@angular/core": "^13.3.11",
+    "@angular/forms": "^13.3.11",
+    "@angular/platform-browser": "^13.3.11",
+    "@angular/platform-browser-dynamic": "^13.3.11",
+    "@angular/router": "^13.3.11",
+    "@ng-bootstrap/ng-bootstrap": "^12.0.0",
+    "@ngx-translate/core": "^14.0.0",
+    "@ngx-translate/http-loader": "^7.0.0",
+    "@stomp/stompjs": "^6.1.2",
     "jquery": "^3.4.1",
     "jquery-ui-dist": "^1.12.1",
-    "papaparse": "^4.6.3",
+    "papaparse": "^5.3.2",
     "pivottable": "^2.23.0",
-    "rxjs": "~6.3.3",
-    "tslib": "^1.9.0",
-    "zone.js": "~0.8.26"
+    "rxjs": "~7.5.7",
+    "tslib": "^2.3.0",
+    "zone.js": "~0.11.8"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.13.0",
-    "@angular/cli": "~7.3.5",
-    "@angular/compiler-cli": "~7.2.0",
-    "@angular/language-service": "~7.2.0",
-    "@biesbjerg/ngx-translate-extract": "^2.3.4",
-    "@types/node": "~8.9.4",
-    "@types/papaparse": "^4.5.9",
-    "codelyzer": "~4.5.0",
+    "@angular-devkit/build-angular": "^13.3.9",
+    "@angular/cli": "^13.3.9",
+    "@angular/compiler-cli": "^13.3.11",
+    "@biesbjerg/ngx-translate-extract": "^7.0.4",
+    "@types/papaparse": "^5.3.5",
     "cpx": "^1.5.0",
-    "ts-node": "~7.0.0",
-    "tslint": "~5.11.0",
-    "typescript": "~3.2.2"
+    "typescript": "~4.6.4"
   }
 }
