@@ -1,1 +1,3 @@
-declare var webkitSpeechRecognition: SpeechRecognition;
+declare var webkitSpeechRecognition: any;
+declare type SpeechRecognitionEvent = any;
+declare type SpeechRecognition = any;
